Attempt to re-initialize video stream when no frames read
We encountered a situation where the cv2 VideoCapture process stops reading frames periodically. Verbose logging of opencv and the FFMPEG backend did not show any errors, and `self.vcap.isOpened()` continued to return `True` even though no frames were read. 

With this PR, we attempt to reinitialize the video stream in this case, and exit if the video stream can't be initialized.